Adding persisted channels and working code
This code takes a static message and pushes it to a IOWorker via a
persistence channel. The IOWorker immediately confirms the message.
For this code path I would assume the messages in journals should get
deleted after a short time interval as message deletion is asynchronous.
But my code demonstrates that the journal space only grows over time.